Moved files and updated docker
